requirements.txt
django==2.0.2
djangorestframework==3.7.7
django-configurations==2.0
pytest==3.4.1
pytest-django==3.1.2
mock==2.0.0
simplejson==3.13.2
django_rest_swagger==2.1.2
factory-boy==2.10.0
